Began modding
There's a little bit of code for the intro, but it isn't quite enough to
be fully working yet. Also, there are a couple more default mods added.

The aliases `xpos`, `ypos`, and `zpos` now exist to move the players.
Use these instead of setting the coordinates directlty on the players.