Terrifying Scroll Effect
This code adds the most terrifying scroll effect possible. It is
implemented with masks, and it is also totally horrible and evil and
scary because of the way it is. I am writing more and more content here
just because I find it cool how vim automatically wraps the words, but
like no other text editor does that because what text editor would just
put enters in your code? Apparently vim if it sees that you're writing a
commit message! :-)